BUG#15956714: REPLICATION DOESN'T WORK IN MYSQL 5.6.80-RC IF STARTED WITH SKIP-INNODB
When a slave was started with --skip-innodb and info file repositories
(default values), replication was incorrectly stopped. If slave is not
doing a migration between info repositories and is using file
repositories, replication can work flawless.

When server is started with info file repositories and it is not doing a
live migration, ignore errors when trying to open table info
repositories.