Bug #14748587 INNODB.INNODB_BUG11766634 FAILING ON DAILY-5.6
The test case innodb_bug11766634.test has been written with the
assumption that innodb_file_per_table is OFF, which was the default
value for versions <= 5.6.5.  But the default value of 
innodb_file_per_table is ON for MySQL versions >= 5.6.6.  So this
configuration variable is explicitly set to OFF in the test case.