ndb - dbacc, duplicated primary keys and failing primary key lookups
Bug #16023068 DUPLICATED PRIMARY KEYS AND FAILING PRIMARY KEY LOOKUPS

In the unique hash indexes some values got wrong reduced hashvalue on expansion
of the internal linear hash table.
This resulted in that key lookups failed since the reduced hashvalue did not match
the hashvalue of the key searched for, and so also insert a duplicate key was
possible under some circumstances.

This patch corrected the calculation of the reduced hashvalue on expansion
by adjusting it for the expansion not yet completed.