ndb - dbacc, calculate the bits needed for quicker exact key operations
this bug should result in extra slow exact key operations for big
tables since it relaxed the condition of valid bits in the reduced
hashvalue stored for each element in the unique hash index.
this makes the reduced hashvalue less selective for filtering out
possible matches in a bucket, so more elements are compared by
value with the key.