BUG#12902514 - MTS RECOVERY TESTS ARE SLOW ON PB (SOLARIS)
A set of tests satisfying the pattern 'rpl_*mts*_crash_safe.test'
was used to perform slow and intolerably slow on solaris.

In fact the bug was largely fixed when the mts-crash-safe tests
were made run through --big-test.
This patch introduces a lighter version of rpl_stm_mixed_mts_rec_crash_safe
and is supposed to run per push.
The current --big-test version set is shorten thanks to
an optimization is done in that STM and MIXED format tests of the mts-crash-safe
set are united into a single one (see renaming and deletion).

Debug print out is activated to hunt over a separate issue of timeouts
that is separted out as Bug 16016262.
The lighter version of rpl_stm_mixed_mts_rec_crash_safe execution on PB should 
provided necessary details.