BUG#13934876 - SHARED MEMORY CONNECTIONS CRASH THE SERVER                              WHEN ENDING
DESCRIPTION:
When mysqld is in shared memory mode under windows and
we stress test the server with a multi threaded client
which also randomly issues "kill connection" command
we see the mysqld process crashing randomly.


ANALYSIS :
When "kill connection xx" is received and executed on an 
arbitrary thread it calls THD::shutdown_active_vio() on the 
THD refered to by connection xx, this results in the 
shared memory being unmapped. However the original thread
serving the connection xx might be in middle of a vio_read
or vio_write and it might end up accessing the unmapped
region of memory resulting in a crash.


FIX :
When "kill connection xx" is executed on an arbitrary thread it calls
THD::shutdown_active_vio() on the THD referred to by xx, this results in the
shared memory being unmapped. However the thread serving the connection xx
might be in middle of a vio_read or vio_write and it might end up accessing
the unmapped region of memory.

All handles are closed and the VIO is cleaned up when vio_delete() is
called and this completes the vio cleanup operation in its entirety.

All references to vio_close has been changed to vio_shutdown to cleanup code
and make its meaning relevant.