This patch fixes bug#13519770 (CODE REVIEW: WRITE_ROW() CRASHES ON INSERT...DUPLICATE KEY UPDATE).
The root cause of this bug is that the return value of handler::get_dup_key
wasn't checked for valid range. It was possible that this method could
return the constant MAX_KEY and subsequent call to key_copy()
could generate a segfault and crash the system. This bug was discovered
during code review and doesn't produce any visible side effect.

We consider the constant MAX_KEY for the value of key index as indicator
for the case when there isn't any information about key that provoked
the duplication error. For such case we require that the storage engine
supports the HA_DUPLICATE_POS flag. If some future storage engine
doesn't support HA_DUPLICATE_POS flag and get_dup_key() returns value MAX_KEY
then the server will be crashed by this assert.
In order to catch the condition mentioned above for mysql server built
without debug we added extra 'else' branch that returns error if the key index
is equal to MAX_KEY.