Bug#16038776: PREPARED STATEMENT IS VERY SLOW WHILE BINLOG WAS ENABLED
Optimization: Substituting the param_markers (the ? bindings
in a prepared statement) for logging is costly, so if the only
log that's open is the binlog, require that we be in SBL mode
as a condition of substitution. If we're in RBL mode, we won't
see the statement, so there's little point in fixing it up.
This holds for DML; DDL will still go through substitution in
case SBL logging is forced for that statement. (Overhead for
dynamic SQL (as opposed to PS with many substitutions) is
comparatively low, anyway.)

Adapted from a submission, extended, and verified by test suite.