BUG#16084426: MYSQLDUMP WITH GTID FAILS TO RESTORE WITH PUZZLING ERROR
In order to restore or provisioning a new server, DBAs can set
@@GLOBAL.GTID_PURGED to a given GTIDs set specifying the transactions
that were imported.
This operation requires that @@GLOBAL.GTID_EXECUTED, and consequently
@@GLOBAL.GTID_PURGED since the latter is a subset of the former, is
empty to be executed, in order to avoid DBAs mistakes that override
server executed GTIDs.
When the requirement was not fulfilled, the following error message was
thrown: "GTID_PURGED can only be set when GTID_EXECUTED is empty.", what
could be misleading since GTID_EXECUTED has global and session scopes.

In order to avoid scope misleading, error messages that mention GTIDs
variables include the variable scope.