BUG#16074161: CRASH IN SLAVE_ROWS_SEARCH_ALGORITHMS_CHECK WHEN SETTING WRONG ALGORITHM
Problem: Setting wrong algorithms for the variable 
SLAVE_ROWS_SEARCH_ALGORITHMS leads to crash, in certain specific
cases. This is due to the fact that the check for values of the
algorithm was not being done as expected.

Fix: We fix this problem by using already existing function
check_not_null() to see if the algorithm values specified is not
null. We also add some additional checks to prevent incorrect
values from being set.