Bug#16056365: UNRECOGNIZED MASTER VERSION MAY LEAD TO CRASH DUE TO DOUBLE MUTEX LOCKING
Problem: In-case of unrecognosed server version,
the mi->data_lock is not unlocked, and the codeflow
jumps to the error section of handle_slave_io() which
tries to unlock it again. This leads to crash in case
the master version is notrecognized or it is very old.

Fix: unlock mi->data_lock in case of error before jumping to
error section.