Bug#11762107 HAVE_MLOCK NEVER SET BY CONFIGURE, AND IF SET BREAKS COMPILATION
Before this fix, mysys source code contained memory locking apis:
- my_malloc_lock
- my_free_lock

The implementation for these api was in fact never built,
as the code had a build break for 64 bit platforms
since it was last changed, around 2007-05-10.

The fix is to simply remove this dead code.

See the bug report for further comments.

Other components may use mlock() and mlockall() directly,
so the cmake checks for HAVE_MLOCK and HAVE_MLOCKALL are preserved.