Bug#14575847: LOG TABLES MAKE ASSUMPTIONS ABOUT CHARSET
Logging to TABLE recorded different queries in some cases
than logging to file did.

The log-table's column was UTF-8, but client's charset
did not factor into conversion; ASCII characters and
byte-sequences that happened to form valid UTF-8
multi-byte characters were copied 1:1, and everything
else was transliterated into escaped hex characters
(which as opposed to entire hex string literals, the
server does not support in its input).
This meant that something that was multiple characters
in the client's charset might be stored verbatim, and
then replayed as a single UTF-8 multi-byte character.
(It would look wrong, even if the byte-sequence itself
was preserved. Casting from UTF-8 to the original charset
would then not render the original sequence.)

The apparently obvious fix (taking into account the
client's charset) is hindered by the fact that queries
may contain multiple encodings (e.g. _koi8r'mySQL').

Therefore, we now store the unmodified query "as is"
(i.e. in a BINARY type MEDIUMBLOB column), which is
always correct.