Bug#14839743: QUERY CACHE INVALIDATED BY DROPPING TEMPORARY TABLE
The problem was that the query cache was needlessly invalidated
if a temporary table is dropped. Since we don't cache queries
using temporary tables and skip queries using temporary tables
during cache lookup, there is no reason to invalidate if only
temporary tables are dropped.

This patch fixes the problem by only invalidating the query cache
during DROP TABLE for the non-temporary tables that are dropped.