Bug#16076275: HANDLER::HA_END_BULK_IN CREATE TABLE .. SELECT, STATEMENT BINLOGGING
Regression from wl#4443.

Due to reorganization of select_create::prepare2()/create_table_from_items.

If locking fails, the temporary table is dropped, but the pointer 'table' is
not reset. So during select_insert::abort_result_set() since table is not NULL
it tries to use the pointer which is no longer valid.

Fixed by resetting table to NULL after it was dropped.