Bug#14586710: INCORRECT RESULTS FOR QUERY WITH NOT IN AND               SUBQUERY THAT EVALUATES TO NULL
When the optimizer transforms a subquery with aggregate 
functions from IN to EXISTS, a HAVING clause that checks if 
fields in the select list of the subquery are NULL is added in
Item_in_subselect::single_value_in_to_exists_transformer().
Later, Item_in_subselect::finalize_exists_transform() 
replaces all the SELECT list fields with Item_int since EXISTS
does not care about the value anyway (and because it makes
EXPLAIN EXTENDED show the somewhat more intuitive 
"... EXISTS (SELECT 1 ...)" instead of e.g.
"... EXISTS (SELECT t1.my_field1, sum(t1.my_field2) ...)".

After this, JOIN::fields_list for the subquery will contain
the Item_ints, not the original Items. 

In the case that the subquery contains no qualifying rows, 
return_zero_rows() will call no_rows_in_result() on all items
in JOIN::fields_list (the items returned from the subquery). 
But in the case of IN-to-EXISTS transformed subqueries, this
list does not contain the Items JOIN::having points to. 
Evaluation of join->having->val_int() in return_zero_rows()
will therefore not be evaluated with the knowledge that there
are zero rows in the result set and thus fails.

The fix is to call no_rows_in_result() on all Items in 
JOIN::all_fields instead of all Items in JOIN::fields_list.