BUG#16105548: MYSQLBINLOG DOES NOT IGNORE HEARTBEAT               EVTS WHEN CONNECTING TO REMOTE SERVER
mysqlbinlog can connect to a remote server and read binary logs
from it. In MySQL 5.6, this tool can even wait for the server to
generate more events and send them. In practice, mysqlbinlog
behaves as a slave connecting to a master. This behavior was
introduced in WL 4783. However, if the server sends an heartbeat,
the tool does not know how to handle it, since it knows nothing
about these events, other than its type code. As a consequence,
the tool bails out and does not read more events from the server.

The fix is to ignore those events that have type code equal to
HEARTBEAT_LOG_EVENT. Adjusting the offset (to make
--stop-position to work identically as if reading from a local
log) is questionable when handling an event of type
heartbeat (since it does not exist on the server's binary
log). As such, this patch just ignores it.