BUG#16170977: MYSQLBINLOG DOES NOT FLUSH LAST EVENT WHEN USING --STOP-NEVER
mysqlbinlog is not flushing the contents of the FILE stream it
uses when dumping the contents of the IO_CACHE after reading
events from the binary log. This means that the last event
processed may not show up immediately in the output when piping
mysqlbinlog output through other stream than stdout.

To fix this we deploy a call to fflush(file) when the IO_CACHE is 
copied to the stream (file) and before reinitializing the 
IO_CACHE.