BUG#16043173 - ASSERTION FAILED: M_STATUS == DA_ERROR
DESCRIPTION:
When the Debug Build server is started with Bin-Logging disabled and "show relaylog events;"
is executed from within a Stored Procedure an assert is triggered.


ANALYSIS :
In mysql_show_relaylog_events() when MasterInfo has not been initialized because
Bin-Logging is disabled, the function fails returning TRUE, however an error is not
set in the Diagnostic area. This results in the ASSERT being triggered in
Diagnostic_area::returned_sqlstate().

The user-visible effect of this bug is that it would appear that SHOW RELAYLOG
EVENTS executed successfully, when it had in fact failed. This would happen 
regardless of if it was used in a stored procedure or not.

FIX :
In mysql_show_relaylog_events() when active_mi is NULL we now call my_error()
instead of my_eof().