Bug#16169676: ASSERT IN FUNCTION ADD_GROUP_AND_DISTINCT_KEYS Bug#16165832: ASSERTION FAILED: JOIN_TAB->CONST_KEYS.               IS_SUBSET(JOIN_TAB->KEYS), FILE SQL_OPTIM
make_join_statistics() may discover new sargables (predicates
that may take advantage of an index) if const tables have been
found. The appropriate index bits are then set in 
JOIN_TAB::const_keys but not in JOIN_TAB::keys. This is incorrect
as const_keys shall be a subset of keys. The fix is to set
the same bits in JT::keys.