Bug#14631798: DEBUG ASSERTION FAILURE IN QUERY CACHE CODE ON SYSBENCH OLTP TEST
The debug mysqld runs additional integrity checks on the query cache.
When checking, don't tell other contestants for the query cache that
we'll throw out the contents at the end of the test, let them wait
and proceed as normal once we're done instead.

The way it was handled could result in one thread running the checks,
and others being told that the cache would be rebuilt by the current
holder of the lock (i.e. the thread running the tests). The other
contestents for the cache would then stop attaching results to their
queries, possibly in medias res, with no attempt at cleaning up the
half finished result sets (seeing as the lock-holder seemed to
promise to discard all of the cash). Seeing as the integrity check
did not in fact discard the query cache, half-finished result sets
might end up in the query cache and later be caught by our sanity
checks.