Bug#16196158 RACE CONDITION IN TABLE PERFORMANCE_SCHEMA.THREADS
Before this fix, the following query:
  SELECT * from performance_schema.threads
could crash, while attempting to read the column PROCESSLIST_INFO

The reason for the failure is a race condition, that can happen
when reading the state of the thread that is changing.

The root cause is that the PFS_lock used to protect dirty reads was used 
incorrectly.

The PROCESSLIST_STATE and PROCESSLIST_INFO columns are protected by
PFS_thread::m_processlist_lock, but the code used PFS_thread::m_lock 
instead.