Move an InnoDB custom unit test from storage/innobase/handler/ha_innodb.cc to unittest/gunit/innodb/ha_innodb-t.cc and convert it to use Google test.
Some of the tests are disabled, because they fail. innobase_convert_name()
has been changed after the tests were developed and the tests were never
run after that. Now the function needs to be fixed or the new results
carefully examined and changed in the test if they are acceptable.