BUG#11745788 - 19947: RETURNING INCONSISTENT ERROR MESSAGES                WHEN ACCESSING NON-EXISTENT OBJECTS
BACKGROUND:
Following are the three cases that describes the way
errors should be reported to the user when user has some grants
or no grants to any particular object in the database.

CASE 1: In general, the rule is followed that if a user tries
to access a non-existent object within his privilege grant
level, we tell the user that the object doesn't exist.

CASE 2:If object is outside of his grant/level we tell him
that, he doesn't have the right to access the object.

CASE 3:If object is outside of his grant level and object
does not exists, we were currently telling the user that
object does not exist, instead of telling that he does not
have the rights to access the object.

NOTE:
Object here refers to a column in this bug context.

ANALYSIS:
The current behavior is of type CASE 3. However the expected
behavior is CASE 2. i.e., The user should not be able to
determine the existence of the column, if he does not
have any rights to that column.

FIX:
This bug is fixed by reporting the error ER_COLUMNACCESS_
DENIED_ERROR instead of ER_BAD_FIELD_ERROR for the columns,
if the user has no rights on the column or when the column
does not exists. The above behavior is similar to the way
errors are reported in case of table/database access.