Bug#14193623 	CLUSTER CRASH - SURVIING NODE FAILURE IN HANDLING LQHKEYCONF FROM RECOVERING NOD
This patch adds some self-checking to packed signals, which hopefully
helps pinpoint the source of corrupt Packed Signals observed in this bug.

Note that this is *not* a fix for this bug, but effectively a 
new assertion.

Both pre-send and post-receive checks are coded, but only the pre-send
checks are activated in non-debug builds, to save CPU.