BUG#14798572: REMOVE UNUSED VARIABLE BINLOG_CAN_BE_CORRUPTED FROM MYSQL_BINLOG_SEND
As part Bug #11747416 A DISK FULL MAKES BINARY LOG CORRUPT,
reading the variable "binlog_can_be_corrupted" was removed
In the existing code the value of this variable is only set,
never read. And also this issue causing compiler warnings.
So the variable is completely redundant and should be removed.

sql/sql_repl.cc:
  Removing dead code