BUG#11745049 - LOG_ERROR VARIABLE IS BLANK WHEN NOT EXPLICITLY                STATED IN MY.CNF
DESCRIPTION:

log_error global variable shows empty string ("") if not
specified at server startup or in configuration file. This
is reported for *nix platform. On windows, it shows
./HOSTNAME.err as its output.

ANALYSIS :

Error output on *nix platforms is by default STDERR. On
Windows it is by default in file, and the default file is
./HOSTNAME.err. If both --log-error and --console is used
on windows, the value of log_error is shown but the output
is directed to stderr.

FIX :

Changed the value of log_error variable to "STDERR" whenever
output is directed to the STDERR.