Bug#16164031: SERVER CRASH IN VIRTUAL HA_ROWS               HANDLER::MULTI_RANGE_READ_INFO_CONST
Consider a query with WHERE conditions:

   ((f1=... OR f2=...) AND (f1=... OR f2=...))
                 OR
   (f1= ...)

   Simplified to: (pred1_a AND pred1_b) OR pred2
   in which pred1_a = "f1=... OR f2=..." etc

The range optimizer now tries to construct range predicates. 
One SEL_TREE with merges representing "pred1_a AND pred1_b" 
and another SEL_TREE without merges representing "pred2" is 
first created. tree_or() is then called on these trees. 
Because the first tree contains merges and the second does 
not, tree_or() calls imerge_list_or_tree() which does this:
  1) perform OR between pred1_a and pred2
  2) make a shallow copy of pred2
  3) perform OR between pred1_b and pred2_cpy

In this bug, the key_or() calls done during step 3 merges 
ranges in pred2_cpy and ends up with a range predicate that 
is always true. The problem is that pred2_cpy is a shallow 
copy so the range predicates created during step 1 are also 
affected. This is the bug.

The fix is to increment the use_count of all range predicates
when creating a shallow copy of a SEL_TREE. By doing so, 
key_or() notices that multiple ranges refer to the same 
SEL_ARG in step 3 and merges ranges in a copy of the SEL_ARG 
instead of the original.