Bug#16222630 STORAGE ENGINE CALL IN JOIN_READ_LAST_KEY() BYPASSES P_S.
Prior to this fix, some table io performed by the mysql server when calling
a storage engine were missing from the statistics collected by the
performance schema.

In particular, the call to the method handler::index_read_last_map()
was not instrumented.

The type of queries for which this call is used are queries executed by the
optimizer by reading the last key of an index.

This happens for example when using an ORDER BY xxx DESC clause.

The root cause is that methods such as handler::index_read_last_map()
should be protected, called only from a public (and instrumented) wrapper such as
handler::ha_index_read_last_map().

This fix:
- changes method handler::index_read_last_map() to be protected.
- adds a public and instrumented method handler::ha_index_read_last_map()
- adjusts caller accordingly.

This resolved the bug initially reported.

This fix also:
- change method handler::index_read_map() to be protected,
  to enforce that existing method handler:ha_index_read_map()
  is used as intended.
- change method handler::index_read_same() to be protected,
  to enforce that existing method handler:ha_index_read_same()
  is used as intended.
- adjusts caller accordingly.

This exposed and resolves a related minor bug in the stored procedure implementation
in lock_db_routines(), which performed some table io on table mysql.proc,
also missing instrumentation for table io in the performance schema.

This second bug had in practice no consequences,
given that the performance schema does not collect table io by default on mysql.* tables.