Bug#16037372 - INNODB PURGE THREADS MAY NEVER WORK IN MYSQL 5.6
Make Purge Coordinator aware about innodb max purge lag setting, if it's set
then keep it as the target for purge threads too.
		        
Note: The bug synopsis overstates the problem. The real issue is that it is
possible that in certain cases the purge coordinator may not use all the
available purge threads.
				          
Approved by Jimmy over IM.