Fix for Bug#14826522 REF ACCESS ON STRING DATA TYPES RETURNS                      NON-MATCHING RECORDS
This bug was caused by the fix for Bug 14682735. The fix for Bug
14682735 was for ref access on string data types and tried to avoid
having the evaluate the ref condition as part of the WHERE condition
after the records were read from the storage engine. This worked in
most cases but there are at least two cases where this caused
non-matching records returned from the storage engine to not be filter
out:

1. If a string field is compared against a string constant that is
   longer than the field, then ref access will be done on a prefix of
   the string constant and can thus return non-matching records. These
   were no longer filtered away.

2. For BINARY/VARBINARY fields with equality against a string constant,
   ref access can return non-matching records even in the case where
   the string constant is shorter than the field. This is due to not
   padding the string to the width of the field when setting up ref
   access.

This patch fixes the regression caused by Bug 14682735 by reverting
most of the changes done in the patch. Thus, after this change we will
again evaluate the ref condition for string data types after the
records are read from the storage engine.