bug#16210351  SLAVE SQL THREAD HANGS WHEN CHECKSUM ERROR HAPPENS AND MTS IS ON
When MTS is on, slave coordinator will hang when encountering a checksum error
on an event from a transaction.

Set coordinater's MTS status to MTS_KILLED_GROUP when reading an event
with a checksum error or an invalid event error if there is at least one
not-terminal event scheduled, so that workers can kill the coordinator
thread and exit normally.