Bug#16210894 SERVER_ID IN INFORMATIONAL MESSAGE WRAPS DUE TO WRONG PRINTF SPECIFIER
Problem:
=======
With a slave having large server-id of (e.g. 4294967291), the
message printed on master when slave starts binlog dump and
log_warnings>1:

[Note] Start binlog_dump to slave_server(-5), pos(test-bin.000002,
107)

Analysis:
========
As per the documentation server_id can be from 0 .. 4294967295.
The server_id is declared as "uint32" type. But during printing
this value the format specifier is given as "%d" which is incorrect.
"%d" is only good enough for signed int output.

Fix:
===
As part of fix "%d" has been replaced with "%u".