ndb - Support large hashmaps in 7.0 and 7.1
Prior this patch the NDB_DEFAULT_HASHMAP_BUCKETS was used both
for number of buckets in new hasmaps and also to determine how much
memory to allocate for hashmap objects and so it also determined the
maximum number of buckets in hashmap supported.

This patch introduce a new NDB_MAX_HASHMAP_BUCKETS(3840) that will be
used when allocating memory for hashmap objects.
NDB_DEFAULT_HASHMAP_BUCKETS(240 or 3840) continues to be used to detemine
the number of buckets in new hashmaps.

Benefits,
1) new 7.0 and 7.1 will be able to cope with >=7.2.7 hashmaps.
   and if nothing else hinder downgrade, downgrade should work

Drawbacks,
1) more memory will be used in both kernel and ndbapi.
   about 7k bytes more per hashmap. in kernel there are preallocated space for
   32 hashmaps so in kernel 230kbytes more static memory will be used.
2) does not apply to old 7.0 and 7.1 releases, so downgrade from 7.2.7 to older
   7.0 and 7.1 will still not work