ndb - use DefaultHashmapSize for default hashmaps
Bug #14800539   MYSQL CLUSTER DOWNGRADE FROM 7.2.8 ( HASHMAP ) TO PREVIOUS VERSIONS IS BROKEN

To make it possible to downgrade from one version with bigger hashmaps to a version with smaller,
a new parameter DefaultHashmapSize is introduced to control the hashmap size used.

To be able to downgrade, this parameter must be set at initial start since it only control
hashmap size for new tables created and there are no way to "reorganize" tables to smaller
hashmaps.

Note, it should be set to the same value for all data and api nodes.

To be compatible with old versions of cluster set this parameter to 240,
otherwise it should be left unset.

A value of 0 (zero) means the same as unset, leaving the decision to what the current version supports by default.