Fix for Bug#15896009 DBT-3 QUERY 19 RUNS SLOWER WITH INDEX CONDITION PUSHDOWN
The cause for this regression is that Q-19 has a fairly complex query
condition. When using index condition pushdown (ICP), we try to push
as much as possible of this condition down to the storage engine. The
ICP code will push down the parts of the query condition that can be
evaluated using fields from the index and fields from records from
earlier tables in the join. In this case ICP is used for the second
table in the join and the index condition that is pushed contains only
references to fields from the first table in the join. This part of
the query condition will (in this case) already have been evaluated
when reading the first table, and since the pushed index condition
does not contain any fields from the index, no records can be filtered
out by the push index condition. Thus, using index condition pushdown
in this case leads to increased CPU usage since we unnecessary
evaluates this condition multiple times extra with no possibility of
filtering out any records.

The fix for this problem is to ensure that we only push conditions to
the storage engine that contain at least one field from the used
index.