Implement proper mysql error code conversions using MysqlErrToSQLStateMap
MySQLConnection DBOperationError first converts the felix error string to the mysql error number
It then converts the error number to the sqlstate and copies the message string from the Error.
The original felix error is then set as the cause of the DBOperationError returned to the user.

MysqlErrToSQLStateMap has two parts: error string to error number; and error number to sqlstate

The error string is returned by felix node driver, e.g. 'ER_DUP_ENTRY'
The error number is the mysql error number, e.g. 1062
The sqlstate is the documented sqlstate corresponding to the error number, e.g. '23000'