WL#6663: Make the MTR binlog suite GTID_MODE agnostic
Due to several problems, the binlog suite is not fully 
gtid-mode=on compliant. These problems range from
the fact that MTR is result file oriented to the fact that
some tests not even have binlog turned on ever.
          
This patch fixes the binlog suite so that in the future 
we might be able to run (most of) it with gtid-mode=on.
The tests left out should be those that don't make sense
with gtid-mode=on.