Fix Bug#16249500 INNODB SILENTLY SWITCHES TO 1 BUFFER POOL INSTANCE IF SIZE IS < 1GB
Note: the following ratiocination assumes the user will not start mysqld
with --innodb-buffer-pool-instances=0. If he does, then we will not emit
a warning to him that the value is adjusted to 1 or 8, depending on the
buffer pool size. This is a limitation in the MySQL config subsystem - we
cannot tell whether the user explicitly set the config knob to its default
value or did not specify it at all in my.cnf or command line.

Exploit the value of 0 as a default value for innodb_buffer_pool_instances.
This way we can tell if the user tried to set it to something != 1 when the
buffer pool size is less than 1 GiB. We use the following logic to set
innodb_buffer_pool_size:

If the size of the BP >= 1 GiB and --innodb-buffer-pool-instances= was not
specified on the command line, and this is not win32, then set to 8, if it
is win32, then set to buffer pool size / 128 MiB.

If the size of the BP < 1 GiB and --innodb-buffer-pool-instances= was not
specified on the command line, then set to 1.

If the size of the BP < 1 GiB and --innodb-buffer-pool-instances=N was
specified on the command line and N != 1, then set to 1 and emit a warning.

Reviewed by:	Kevin, Inaam (rb:1984)